# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/checksymptomss": {
            "get": {
                "description": "list checksymptoms entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List checksymptoms entities",
                "operationId": "list-checksymptoms",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Checksymptoms"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create checksymptoms",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create checksymptoms",
                "operationId": "create-checksymptoms",
                "parameters": [
                    {
                        "description": "Checksymptoms entity",
                        "name": "checksymptoms",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Checksymptoms"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Checksymptoms"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/diseases": {
            "get": {
                "description": "list disease entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List disease entities",
                "operationId": "list-disease",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Disease"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create disease",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create disease",
                "operationId": "create-disease",
                "parameters": [
                    {
                        "description": "disease entity",
                        "name": "disease",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Disease"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Disease"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/diseases/{id}": {
            "get": {
                "description": "get disease by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a disease entity by ID",
                "operationId": "get-disease",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Disease ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Disease"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update disease by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a disease entity by ID",
                "operationId": "update-disease",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Disease ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Disease entity",
                        "name": "disease",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Disease"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Disease"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get disease by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a disease entity by ID",
                "operationId": "delete-disease",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Disease ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.Checksymptoms": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "disease": {
                    "type": "integer"
                },
                "doctorOrderSheet": {
                    "type": "integer"
                },
                "patient": {
                    "type": "integer"
                },
                "personnel": {
                    "type": "integer"
                }
            }
        },
...
        "ent.Checksymptoms": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "Date holds the value of the \"date\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ChecksymptomsQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ChecksymptomsEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.ChecksymptomsEdges": {
            "type": "object",
            "properties": {
                "disease": {
                    "description": "Disease holds the value of the disease edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Disease"
                },
                "doctorordersheet": {
                    "description": "Doctorordersheet holds the value of the doctorordersheet edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DoctorOrderSheet"
                },
                "patient": {
                    "description": "Patient holds the value of the patient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patient"
                },
                "personnel": {
                    "description": "Personnel holds the value of the personnel edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Personnel"
                }
            }
        },
        "ent.Dentalappointment": {
            "type": "object",
            "properties": {
                "appointtime": {
                    "description": "Appointtime holds the value of the \"appointtime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DentalappointmentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DentalappointmentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Disease": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DiseaseQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DiseaseEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
        "ent.DiseaseEdges": {
            "type": "object",
            "properties": {
                "checksymptoms": {
                    "description": "Checksymptoms holds the value of the Checksymptoms edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Checksymptoms"
                    }
                }
            }
        },
        "ent.DoctorOrderSheet": {
            "type": "object",
            "properties": {
                "Name": {
                    "description": "Name holds the value of the \"Name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DoctorOrderSheetQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DoctorOrderSheetEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "note": {
                    "description": "Note holds the value of the \"note\" field.",
                    "type": "string"
                },
                "time": {
                    "description": "Time holds the value of the \"time\" field.",
                    "type": "string"
                }
            }
        },
        "ent.DoctorOrderSheetEdges": {
            "type": "object",
            "properties": {
                "checksymptoms": {
                    "description": "Checksymptoms holds the value of the Checksymptoms edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Checksymptoms"
                    }
                }
            }
        },
        "ent.Patient": {
            "type": "object",
            "properties": {
                "birthday": {
                    "description": "Birthday holds the value of the \"birthday\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PatientQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientEdges"
                },
                "gender": {
                    "description": "Gender holds the value of the \"gender\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
        "ent.PatientEdges": {
            "type": "object",
            "properties": {
                "antenatalinformation": {
                    "description": "Antenatalinformation holds the value of the Antenatalinformation edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Antenatalinformation"
                    }
                },
                "bonedisease": {
                    "description": "Bonedisease holds the value of the Bonedisease edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Bonedisease"
                    }
                },
                "checksymptoms": {
                    "description": "Checksymptoms holds the value of the Checksymptoms edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Checksymptoms"
                    }
                },
                "dentalappointment": {
                    "description": "Dentalappointment holds the value of the Dentalappointment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Dentalappointment"
                    }
                },
                "physicaltherapyrecord": {
                    "description": "Physicaltherapyrecord holds the value of the physicaltherapyrecord edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Physicaltherapyrecord"
                    }
                },
                "surgeryappointment": {
                    "description": "Surgeryappointment holds the value of the Surgeryappointment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Surgeryappointment"
                    }
                }
            }
        },
        "ent.Personnel": {
            "type": "object",
            "properties": {
                "department": {
                    "description": "Department holds the value of the \"department\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PersonnelQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PersonnelEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "password": {
                    "description": "Password holds the value of the \"password\" field.",
                    "type": "string"
                },
                "user": {
                    "description": "User holds the value of the \"user\" field.",
                    "type": "string"
                }
            }
        },
        "ent.PersonnelEdges": {
            "type": "object",
            "properties": {
                "antenatalinformation": {
                    "description": "Antenatalinformation holds the value of the Antenatalinformation edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Antenatalinformation"
                    }
                },
                "bonedisease": {
                    "description": "Bonedisease holds the value of the Bonedisease edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Bonedisease"
                    }
                },
                "checksymptoms": {
                    "description": "Checksymptoms holds the value of the Checksymptoms edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Checksymptoms"
                    }
                },
                "dentalappointment": {
                    "description": "Dentalappointment holds the value of the Dentalappointment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Dentalappointment"
                    }
                },
                "physicaltherapyrecord": {
                    "description": "Physicaltherapyrecord holds the value of the physicaltherapyrecord edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Physicaltherapyrecord"
                    }
                },
                "surgeryappointment": {
                    "description": "Surgeryappointment holds the value of the Surgeryappointment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Surgeryappointment"
                    }
                }
            }
        },
...
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "authorizationUrl": "",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "authorizationUrl": "",
...
}